Task provider separated into helper class
