+++ soundProvider/pcm.cpp
@@ -14,25 +14,8 @@ unsigned long int SoundProviderPcm::getFrequency() {
 	return freq;
 }
 
-void SoundProviderPcm::provider_start() {
-	if (taskHandle == NULL) {
-		xTaskCreate(reinterpret_cast<TaskFunction_t>(&SoundProviderPcm::taskCode), "SProviderPcm", 512, this, 10, &taskHandle);
-	}
-}
-
-void SoundProviderPcm::provider_stop() {
-	if (taskHandle != NULL) {
-		vTaskDelete(taskHandle);
-		taskHandle = NULL;
-	}
-}
-
-void SoundProviderPcm::taskCode() {
+void SoundProviderPcm::task_code() {
 	for (SoundPos i = 0; i < len; i++) {
 		postSample(data[i]);
 	}
-	postControl(END);
-	TaskHandle_t handle = taskHandle;
-	taskHandle = NULL;
-	vTaskDelete(handle);
 }
